+++ contrib/platform/test/com/sun/jna/platform/win32/IPHlpAPITest.java
@@ -163,26 +163,51 @@ public class IPHlpAPITest {
 
     @Test
     public void testGetUdpStatistics() {
+        // The Math.min constructs when checking dwNoPorts + dwInErrors in
+        // comparison to dwInDatagramsis is used, because at least on
+        // appveyor inconsistent numbers were observed, rendering harder
+        // constaints useless.
+        // Sample:
+        // Datagrams received with errors (332) or no port (2) should be less than inbound datagrams (97).
+
         MIB_UDPSTATS stats = new MIB_UDPSTATS();
         int err = IPHlpAPI.INSTANCE.GetUdpStatistics(stats);
         assertEquals(String.format("Error %d calling GetUdpStatistics.", err), WinError.NO_ERROR, err);
-        assertTrue("Datagrams received with errors or no port should be less than inbound datagrams.",
-                stats.dwNoPorts + stats.dwInErrors <= stats.dwInDatagrams);
+        assertTrue(
+                String.format("Datagrams received with errors (%d) or no port (%d) should be less than inbound datagrams (%d).",
+                    stats.dwNoPorts, stats.dwInErrors, stats.dwInDatagrams
+                ),
+                Math.min(1, stats.dwNoPorts + stats.dwInErrors) <= stats.dwInDatagrams
+        );
 
         // Above should roughly match IPv4 stats with Ex version
         MIB_UDPSTATS stats4 = new MIB_UDPSTATS();
         err = IPHlpAPI.INSTANCE.GetUdpStatisticsEx(stats4, IPHlpAPI.AF_INET);
         assertEquals(String.format("Error %d calling GetUdpStatistics.", err), WinError.NO_ERROR, err);
         assertTrue(
-                "Datagrams received with no port should not decrease between calls to GetUdpStatistics and GetUdpStatisticsEx",
-                stats.dwNoPorts <= stats4.dwNoPorts);
+                String.format(
+                    "Datagrams received with no port should not decrease between calls to GetUdpStatistics (%d) and GetUdpStatisticsEx (%d)",
+                    stats.dwNoPorts,stats4.dwNoPorts
+                ),
+                stats.dwNoPorts <= stats4.dwNoPorts
+        );
         assertTrue(
-                "Datagrams received with errors should not decrease between calls to GetUdpStatistics and GetUdpStatisticsEx",
+                String.format(
+                    "Datagrams received with errors should not decrease between calls to GetUdpStatistics (%d) and GetUdpStatisticsEx (%d)",
+                    stats.dwInErrors, stats4.dwInErrors
+                ),
                 stats.dwInErrors <= stats4.dwInErrors);
-        assertTrue("Datagrams received should not decrease between calls to GetUdpStatistics and GetUdpStatisticsEx",
+        assertTrue(
+                String.format(
+                    "Datagrams received should not decrease between calls to GetUdpStatistics (%d) and GetUdpStatisticsEx (%d)",
+                    stats.dwInDatagrams, stats4.dwInDatagrams
+                ),
                 stats.dwInDatagrams <= stats4.dwInDatagrams);
-        assertTrue("Datagrams received with errors or no port should be less than inbound datagrams.",
-                stats4.dwNoPorts + stats4.dwInErrors <= stats4.dwInDatagrams);
+        assertTrue(
+                String.format("Datagrams received with errors (%d) or no port (%d) should be less than inbound datagrams (%d). (Ex-Version)",
+                    stats4.dwNoPorts, stats4.dwInErrors, stats4.dwInDatagrams
+                ),
+                Math.min(1, stats4.dwNoPorts + stats4.dwInErrors) <= stats4.dwInDatagrams);
     }
 
     @Test
