+++ contrib/platform/test/com/sun/jna/platform/win32/Secur32Test.java
@@ -190,84 +190,6 @@ public class Secur32Test extends TestCase {
                 phClientCredential));
     }
 
-    public void testImpersonateRevertSecurityContext() {
-        // client ----------- acquire outbound credential handle
-        CredHandle phClientCredential = new CredHandle();
-        TimeStamp ptsClientExpiry = new TimeStamp();
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.AcquireCredentialsHandle(
-                null, "Negotiate", Sspi.SECPKG_CRED_OUTBOUND, null, null, null,
-                null, phClientCredential, ptsClientExpiry));
-        // client ----------- security context
-        CtxtHandle phClientContext = new CtxtHandle();
-        IntByReference pfClientContextAttr = new IntByReference();
-        // server ----------- acquire inbound credential handle
-        CredHandle phServerCredential = new CredHandle();
-        TimeStamp ptsServerExpiry = new TimeStamp();
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.AcquireCredentialsHandle(
-                null, "Negotiate", Sspi.SECPKG_CRED_INBOUND, null, null, null,
-                null, phServerCredential, ptsServerExpiry));
-        // server ----------- security context
-        CtxtHandle phServerContext = new CtxtHandle();
-        ManagedSecBufferDesc pbServerToken = null;
-        IntByReference pfServerContextAttr = new IntByReference();
-        int clientRc = W32Errors.SEC_I_CONTINUE_NEEDED;
-        int serverRc = W32Errors.SEC_I_CONTINUE_NEEDED;
-        do {
-            // client ----------- initialize security context, produce a client token
-            // client token returned is always new
-            ManagedSecBufferDesc pbClientToken = new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN, Sspi.MAX_TOKEN_SIZE);
-            if (clientRc == W32Errors.SEC_I_CONTINUE_NEEDED) {
-                // server token is empty the first time
-                ManagedSecBufferDesc pbServerTokenCopy = pbServerToken == null
-                        ? null : new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN, pbServerToken.getBuffer(0).getBytes());
-                clientRc = Secur32.INSTANCE.InitializeSecurityContext(
-                        phClientCredential,
-                        phClientContext.isNull() ? null : phClientContext,
-                        Advapi32Util.getUserName(),
-                        Sspi.ISC_REQ_CONNECTION,
-                        0,
-                        Sspi.SECURITY_NATIVE_DREP,
-                        pbServerTokenCopy,
-                        0,
-                        phClientContext,
-                        pbClientToken,
-                        pfClientContextAttr,
-                        null);
-                assertTrue(clientRc == W32Errors.SEC_I_CONTINUE_NEEDED || clientRc == W32Errors.SEC_E_OK);
-            }
-            // server ----------- accept security context, produce a server token
-            if (serverRc == W32Errors.SEC_I_CONTINUE_NEEDED) {
-                pbServerToken = new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN, Sspi.MAX_TOKEN_SIZE);
-                ManagedSecBufferDesc pbClientTokenByValue = new ManagedSecBufferDesc(Sspi.SECBUFFER_TOKEN, pbClientToken.getBuffer(0).getBytes());
-                serverRc = Secur32.INSTANCE.AcceptSecurityContext(phServerCredential,
-                        phServerContext.isNull() ? null : phServerContext,
-                        pbClientTokenByValue,
-                        Sspi.ISC_REQ_CONNECTION,
-                        Sspi.SECURITY_NATIVE_DREP,
-                        phServerContext,
-                        pbServerToken,
-                        pfServerContextAttr,
-                        ptsServerExpiry);
-                assertTrue(serverRc == W32Errors.SEC_I_CONTINUE_NEEDED || serverRc == W32Errors.SEC_E_OK);
-            }
-        } while (serverRc != W32Errors.SEC_E_OK || clientRc != W32Errors.SEC_E_OK);
-        // impersonate
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.ImpersonateSecurityContext(
-                phServerContext));
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.RevertSecurityContext(
-                phServerContext));
-        // release server context
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.DeleteSecurityContext(
-                phServerContext));
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.FreeCredentialsHandle(
-                phServerCredential));
-        // release client context
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.DeleteSecurityContext(
-                phClientContext));
-        assertEquals(W32Errors.SEC_E_OK, Secur32.INSTANCE.FreeCredentialsHandle(
-                phClientCredential));
-    }
-
     public void testEnumerateSecurityPackages() {
         IntByReference pcPackages = new IntByReference();
         PSecPkgInfo.ByReference pPackageInfo = new PSecPkgInfo.ByReference();
